- testing scripts update (gen synth and real world) - rmat graph is default now - sssp is integer now - create vgl graph prints extra time stats - ".graph" extension for files

+++ apps/sssp/sssp.cpp
@@ -52,12 +52,12 @@ int main(int argc, const char * argv[])
         // do calculations
         cout << "Computations started..." << endl;
         cout << "Doing " << parser.get_number_of_rounds() << " SSSP iterations..." << endl;
-        EdgesArray_Vect<float> weights(graph);
+        EdgesArray_Vect<int> weights(graph);
         weights.set_all_random(MAX_WEIGHT);
         for(int i = 0; i < parser.get_number_of_rounds(); i++)
         {
             int source_vertex = graph.select_random_vertex(ORIGINAL);
-            VerticesArray<float> distances(graph, Parser::convert_traversal_type(parser.get_traversal_direction()));
+            VerticesArray<int> distances(graph, Parser::convert_traversal_type(parser.get_traversal_direction()));
 
             performance_stats.reset_timers();
             ShortestPaths::nec_dijkstra(graph, weights, distances, source_vertex,
@@ -68,7 +68,7 @@ int main(int argc, const char * argv[])
             // check if required
             if(parser.get_check_flag())
             {
-                VerticesArray<float> check_distances(graph, SCATTER);
+                VerticesArray<int> check_distances(graph, SCATTER);
                 ShortestPaths::seq_dijkstra(graph, weights, check_distances, source_vertex);
                 verify_results(distances, check_distances);
             }
